Update homework_02.ipynb
Probably just a typo, did you mean to write this?
What confuses me a bit is the use of `x.lower()`. As I understand we want column of tokens (not the string itself), but calling `lower()` on a `Token`-object returns an integer.
Due to this and the use of `x.lemma_` instead of `x.lemma` (which explicitly returns the lower-cased text) got me wondering whether we might also be collecting upper-cased tokens but only print lower-cased text or something else?
I'd appreciate if you could elaborate on that.